arm build

+++ lib/controllers/src/ServoMotor.cpp
@@ -1,15 +1,15 @@
-#include "../include/ServoMotor.h"
+#include "ServoMotor.h" //TODO(qinyang-bao): fix all compile errors and renable servo code
 
 #include <cmath>
 
-ServoMotor::ServoMotor(PinName pwm, bool inverted, float min_pulsewidth_ms, float max_pulsewidth_ms, float limit)
+ServoMotor::ServoMotor(PinName pwm, bool inverted, double min_pulsewidth_ms, double max_pulsewidth_ms, double limit)
     : Motor(NC, NC),
       m_servo(pwm, Servo::CONT_SERVO, limit, max_pulsewidth_ms, min_pulsewidth_ms),
       m_inverted(inverted){};
 
 ServoMotor::ServoMotor(t_motorConfig motorConfig) : ServoMotor(motorConfig.pwmPin, motorConfig.inverted) {}
 
-void ServoMotor::setPower(float percentage) {
+void ServoMotor::setPower(double percentage) {
   m_servo.set_speed(percentage * (m_inverted ? -1.0 : +1.0));
 }
 
@@ -18,6 +18,6 @@ ServoMotor& ServoMotor::operator=(int percentage) {
   return *this;
 }
 
-float ServoMotor::getPower() {
+double ServoMotor::getPower() {
   return m_servo.read();
 }
